// apps/react-frontend/src/api/client/models/create-container-transport-request.ts
export interface CreateContainerTransportRequest {
    /**
     * 
     * @type {string}
     * @memberof CreateContainerTransportRequest
     */
    'consignorPartyId': string;
    /**
     * 
     * @type {string}
     * @memberof CreateContainerTransportRequest
     */
    'pickupDateTime': string;
    /**
     * 
     * @type {string}
     * @memberof CreateContainerTransportRequest
     */
    'deliveryDateTime'?: string;
    /**
     * 
     * @type {string}
     * @memberof CreateContainerTransportRequest
     */
    'transportType': CreateContainerTransportRequestTransportTypeEnum;
    /**
     * 
     * @type {string}
     * @memberof CreateContainerTransportRequest
     */
    'containerOperation': CreateContainerTransportRequestContainerOperationEnum;
    /**
     * 
     * @type {string}
     * @memberof CreateContainerTransportRequest
     */
    'driverId'?: string;
    /**
     * 
     * @type {string}
     * @memberof CreateContainerTransportRequest
     */
    'carrierPartyId': string;
    /**
     * 
     * @type {string}
     * @memberof CreateContainerTransportRequest
     */
    'pickupCompanyId'?: string;
    /**
     * 
     * @type {string}
     * @memberof CreateContainerTransportRequest
     */
    'pickupCompanyBranchId'?: string;
    /**
     * 
     * @type {string}
     * @memberof CreateContainerTransportRequest
     */
    'pickupStreet': string;
    /**
     * 
     * @type {string}
     * @memberof CreateContainerTransportRequest
     */
    'pickupBuildingNumber': string;
    /**
     * 
     * @type {string}
     * @memberof CreateContainerTransportRequest
     */
    'pickupPostalCode': string;
    /**
     * 
     * @type {string}
     * @memberof CreateContainerTransportRequest
     */
    'pickupCity': string;
    /**
     * 
     * @type {string}
     * @memberof CreateContainerTransportRequest
     */
    'deliveryCompanyId'?: string;
    /**
     * 
     * @type {string}
     * @memberof CreateContainerTransportRequest
     */
    'deliveryCompanyBranchId'?: string;
    /**
     * 
     * @type {string}
     * @memberof CreateContainerTransportRequest
     */
    'deliveryStreet': string;
    /**
     * 
     * @type {string}
     * @memberof CreateContainerTransportRequest
     */
    'deliveryBuildingNumber': string;
    /**
     * 
     * @type {string}
     * @memberof CreateContainerTransportRequest
     */
    'deliveryPostalCode': string;
    /**
     * 
     * @type {string}
     * @memberof CreateContainerTransportRequest
     */
    'deliveryCity': string;
    /**
     * 
     * @type {string}
     * @memberof CreateContainerTransportRequest
     */
    'truckId'?: string;
    /**
     * 
     * @type {string}
     * @memberof CreateContainerTransportRequest
     */
    'containerId'?: string;
    /**
     * 
     * @type {string}
     * @memberof CreateContainerTransportRequest
     */
    'note': string;
}

export const CreateContainerTransportRequestTransportTypeEnum = {
    Container: 'CONTAINER',
    Waste: 'WASTE'
} as const;

export type CreateContainerTransportRequestTransportTypeEnum = typeof CreateContainerTransportRequestTransportTypeEnum[keyof typeof CreateContainerTransportRequestTransportTypeEnum];
export const CreateContainerTransportRequestContainerOperationEnum = {
    Exchange: 'EXCHANGE',
    Pickup: 'PICKUP',
    Empty: 'EMPTY',
    Delivery: 'DELIVERY'
} as const;

export type CreateContainerTransportRequestContainerOperationEnum = typeof CreateContainerTransportRequestContainerOperationEnum[keyof typeof CreateContainerTransportRequestContainerOperationEnum];