// apps/react-frontend/src/api/client/models/create-waste-transport-request.ts
export interface CreateWasteTransportRequest {
    /**
     * 
     * @type {string}
     * @memberof CreateWasteTransportRequest
     */
    'consigneePartyId': string;
    /**
     * 
     * @type {string}
     * @memberof CreateWasteTransportRequest
     */
    'pickupPartyId': string;
    /**
     * 
     * @type {string}
     * @memberof CreateWasteTransportRequest
     */
    'consignorPartyId': string;
    /**
     * 
     * @type {string}
     * @memberof CreateWasteTransportRequest
     */
    'pickupDateTime': string;
    /**
     * 
     * @type {string}
     * @memberof CreateWasteTransportRequest
     */
    'deliveryDateTime'?: string;
    /**
     * 
     * @type {string}
     * @memberof CreateWasteTransportRequest
     */
    'containerOperation': CreateWasteTransportRequestContainerOperationEnum;
    /**
     * 
     * @type {string}
     * @memberof CreateWasteTransportRequest
     */
    'transportType': CreateWasteTransportRequestTransportTypeEnum;
    /**
     * 
     * @type {string}
     * @memberof CreateWasteTransportRequest
     */
    'driverId'?: string;
    /**
     * 
     * @type {string}
     * @memberof CreateWasteTransportRequest
     */
    'carrierPartyId': string;
    /**
     * 
     * @type {string}
     * @memberof CreateWasteTransportRequest
     */
    'pickupCompanyId'?: string;
    /**
     * 
     * @type {string}
     * @memberof CreateWasteTransportRequest
     */
    'pickupCompanyBranchId'?: string;
    /**
     * 
     * @type {string}
     * @memberof CreateWasteTransportRequest
     */
    'pickupStreet': string;
    /**
     * 
     * @type {string}
     * @memberof CreateWasteTransportRequest
     */
    'pickupBuildingNumber': string;
    /**
     * 
     * @type {string}
     * @memberof CreateWasteTransportRequest
     */
    'pickupPostalCode': string;
    /**
     * 
     * @type {string}
     * @memberof CreateWasteTransportRequest
     */
    'pickupCity': string;
    /**
     * 
     * @type {string}
     * @memberof CreateWasteTransportRequest
     */
    'deliveryCompanyId'?: string;
    /**
     * 
     * @type {string}
     * @memberof CreateWasteTransportRequest
     */
    'deliveryCompanyBranchId'?: string;
    /**
     * 
     * @type {string}
     * @memberof CreateWasteTransportRequest
     */
    'deliveryStreet': string;
    /**
     * 
     * @type {string}
     * @memberof CreateWasteTransportRequest
     */
    'deliveryBuildingNumber': string;
    /**
     * 
     * @type {string}
     * @memberof CreateWasteTransportRequest
     */
    'deliveryPostalCode': string;
    /**
     * 
     * @type {string}
     * @memberof CreateWasteTransportRequest
     */
    'deliveryCity': string;
    /**
     * 
     * @type {string}
     * @memberof CreateWasteTransportRequest
     */
    'truckId'?: string;
    /**
     * 
     * @type {string}
     * @memberof CreateWasteTransportRequest
     */
    'containerId'?: string;
    /**
     * 
     * @type {string}
     * @memberof CreateWasteTransportRequest
     */
    'note': string;
    /**
     * 
     * @type {number}
     * @memberof CreateWasteTransportRequest
     */
    'consignorClassification': number;
    /**
     * 
     * @type {string}
     * @memberof CreateWasteTransportRequest
     */
    'wasteStreamNumber'?: string;
    /**
     * 
     * @type {number}
     * @memberof CreateWasteTransportRequest
     */
    'weight': number;
    /**
     * 
     * @type {string}
     * @memberof CreateWasteTransportRequest
     */
    'unit': string;
    /**
     * 
     * @type {number}
     * @memberof CreateWasteTransportRequest
     */
    'quantity': number;
    /**
     * 
     * @type {string}
     * @memberof CreateWasteTransportRequest
     */
    'goodsName': string;
    /**
     * 
     * @type {string}
     * @memberof CreateWasteTransportRequest
     */
    'euralCode': string;
    /**
     * 
     * @type {string}
     * @memberof CreateWasteTransportRequest
     */
    'processingMethodCode': string;
}

export const CreateWasteTransportRequestContainerOperationEnum = {
    Exchange: 'EXCHANGE',
    Pickup: 'PICKUP',
    Empty: 'EMPTY',
    Delivery: 'DELIVERY'
} as const;

export type CreateWasteTransportRequestContainerOperationEnum = typeof CreateWasteTransportRequestContainerOperationEnum[keyof typeof CreateWasteTransportRequestContainerOperationEnum];
export const CreateWasteTransportRequestTransportTypeEnum = {
    Container: 'CONTAINER',
    Waste: 'WASTE'
} as const;

export type CreateWasteTransportRequestTransportTypeEnum = typeof CreateWasteTransportRequestTransportTypeEnum[keyof typeof CreateWasteTransportRequestTransportTypeEnum];